literal fixes

+++ TheiaLang/Parser.cs
@@ -789,7 +789,7 @@ public class Parser(List<Token> tokens)
         {
             (object Value, string Type, TypeKind kind, BuiltinType? builtinType) lit = Previous().Lexeme switch
             {
-                string s when int.TryParse(s, out int i)       => (i, "int",   Scalar, INT),
+                string s when long.TryParse(s, out long i)     => (i, "int",   Scalar, INT),
                 string s when double.TryParse(s, out double d) => (d, "float", Scalar, FLOAT),
                 string s when bool.TryParse(s, out bool b)     => (b, "bool",  Scalar, BOOL),
                 _ => throw new Exception("Invalid literal")
